+++ CHANGELOG.md
@@ -6,16 +6,19 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - BREAKING: STFU rework ([#146]).
+- BREAKING: regoRuleReference in config now optional ([#188]).
 - Version now a String instead of enum ([#156]).
 - `operator-rs` `0.6.0` → `0.8.0` ([#177]).
 - Custom resource example now points to regorule-operator service ([#177]).
+- `snafu` `0.6.0` → `0.7.0` ([#188]).
 
 ### Removed 
-- Configurable Port from code and product config ([#156]).
+- Configurable Port from code and product config ([#188]).
 
 [#146]: https://github.com/stackabletech/opa-operator/pull/146
 [#156]: https://github.com/stackabletech/opa-operator/pull/156
 [#177]: https://github.com/stackabletech/opa-operator/pull/177
+[#188]: https://github.com/stackabletech/opa-operator/pull/188
 
 ## [0.6.0] - 2021-12-06